baseplates/requirements.txt
pillow==10.4.0
PyYAML==6.0.1
